mycart/requirements.txt
asgiref==3.5.2
auditlog3==1.0.1
Django==3.2.13
django-environ==0.9.0
django-filter==22.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
